fix: turn sitemap.ts metadata file to route.ts
